requirements.txt
annotated-types==0.6.0
build==1.0.3
CacheControl==0.13.1
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.7
dulwich==0.21.6
fastjsonschema==2.19.0
filelock==3.13.1
idna==3.6
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyproject_hooks==1.0.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
tomlkit==0.12.3
trove-classifiers==2023.11.29
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.24.7
xattr==0.10.1
zipp==3.17.0

fastapi~=0.101.1
sqlmodel~=0.0.12
uvicorn~=0.23.2
jwt~=1.3.1
django-qrcode~=0.3
google~=3.0.0
starlette~=0.27.0
qrcode~=7.4.2
lxml~=4.9.3
pyotp~=2.9.0
psycopg2~=2.9.9
SQLAlchemy~=2.0.23